Move mocks to test dir
