+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Pytest"""
+"""Pytest."""
